requirements.txt
Flask==3.1.2
Werkzeug==3.1.5
mysql-connector-python==9.6.0
python-dotenv==1.2.1
gunicorn==25.1.0